requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
openai
python-dotenv
mlflow
loguru
pytest
requests
pandas
langchain
langchain-text-splitters
faiss-cpu
transformers
torch
PyYAML
tqdm
huggingface-hub
pdfplumber
accelerate
pytesseract 
pdf2image
streamlit
huggingface-hub